+++ CarCenterBusinessLogic/ViewModels/CarKitViewModel.cs
@@ -9,7 +9,7 @@ namespace CarCenterBusinessLogic.ViewModels
         public int KitId { set; get; }
         public string KitName { set; get; }
         public string CarName { set; get; }
-        public int KitCost { set; get; }
+        public decimal KitCost { set; get; }
         public int KitCount { set; get; }
         public DateTime InstallationDate { set; get; }
     }
